Fixed structure of `-codebook-extractor` project.
Basically just Maven-ified it (mostly).

Still a lot to do:

* Doesn't build yet, due to character set issues that need to be fixed.
* Refactor packages a bit.
* Apply Eclipse formatter.
* Wire it up to produce the XML files as part of the Maven build.

CBBI-343
