+++ bluebutton-data-model-codebook-extractor/.gitignore
@@ -1,0 +1,7 @@
+# Ignore Eclipse metadata files, but include .gitignore and .gitattributes.
+/.*
+!/.gitignore
+!/.gitattributes
+
+# Ignore Maven output folders
+/target*/
