--- bluebutton-data-model-codebook-extractor/.settings/org.jboss.ide.eclipse.as.core.prefs
@@ -1,2 +1,0 @@
-eclipse.preferences.version=1
-org.jboss.ide.eclipse.as.core.singledeployable.deployableList=
