--- bluebutton-data-model-codebook-extractor/src/gov/hhs/cms/bluebutton/data/codebook/util/XmlUtil.java
@@ -1,47 +1,0 @@
-package gov.hhs.cms.bluebutton.data.codebook.util;
-
-import java.io.StringReader;
-
-import javax.xml.parsers.DocumentBuilderFactory;
-
-import org.w3c.dom.Node;
-import org.w3c.dom.bootstrap.DOMImplementationRegistry;
-import org.w3c.dom.ls.DOMImplementationLS;
-import org.w3c.dom.ls.LSSerializer;
-import org.xml.sax.InputSource;
-
-public class XmlUtil {
-	
-	
-	/**
-	 * Formats an XML string for pretty printing. Requires Java 1.6.
-	 * 
-	 * @author Cassidy.Shay
-	 * @param xml
-	 * @return pretty-print formatted XML
-	 */
-	public static String prettyPrintXml(String xml)
-	{
-		try
-		{
-			final InputSource src = new InputSource(new StringReader(xml));
-			final Node document = DocumentBuilderFactory.newInstance().newDocumentBuilder().parse(src).getDocumentElement();
-			final Boolean keepDeclaration = Boolean.valueOf(xml.startsWith("<?xml"));
-
-			//May need this: System.setProperty(DOMImplementationRegistry.PROPERTY,"com.sun.org.apache.xerces.internal.dom.DOMImplementationSourceImpl");
-
-			final DOMImplementationRegistry registry = DOMImplementationRegistry.newInstance();
-			final DOMImplementationLS impl = (DOMImplementationLS) registry.getDOMImplementation("LS");
-			final LSSerializer writer = impl.createLSSerializer();
-
-			writer.getDomConfig().setParameter("format-pretty-print", Boolean.TRUE); // Set this to true if the output needs to be beautified.
-			writer.getDomConfig().setParameter("xml-declaration", keepDeclaration); // Set this to true if the declaration is needed to be outputted.
-
-			return writer.writeToString(document);
-		}
-		catch (Exception e)
-		{
-			throw new RuntimeException(e);
-		}
-	}
-}
