+++ pom.xml
@@ -50,6 +50,7 @@
 		<module>bluebutton-data-model-codegen</module>
 		<module>bluebutton-data-model-rif</module>
 		<module>bluebutton-data-model-rif-samples</module>
+		<module>bluebutton-data-model-codebook-extractor</module>
 	</modules>
 
 	<properties>
